fix(suse-initrd): exclude modules that are built-in
When checking the required modules a certain module needs, we need
to check whether those are already built-in or not.
If they are, we exclude them from the list of modules to install.

This is necessary as otherwise dracut-install fails when stumbling
over modules thar are already built-in.

Signed-off-by: Oscar Salvador <osalvador@suse.de>
(cherry picked from commit d393f006fa8d991465f879e78371ab8864d86058)