feat(nvmf): add NBFT connect support, and fix configuration priority
Connect to NVMeoF targets according to the following priority list:

 1 FC autoconnect from kernel commandline
 2 NBFT
 3 discovery.conf
 4 FC autoconnect without kernel command line

This is the same priority list as before, except that it's more explicit
and that the NBFT has been added in position 2. This means that the NBFT
takes precedence over cmdline.d arguments, which is necessary in order
to make sure that the system honours changed firmware parameters.

In cases where this behavior is undesirable (like recovery in the presence of a
broken NBFT, the user can disable NBFT boot using the parameter "rd.nvmf.nonbft".

jsc#PED-1449

(cherry picked from commit b8b8b0870bb887773485fa0a8eb0ccc51ef11a4b)