chore(suse): change default persistent policy
When dracut generates the initramfs, it must refer to disks and partitions to
be mounted in a persistent manner, to make sure the system will boot
correctly. By default, dracut uses /dev/mapper device names.
For example, when dracut detects multipath devices, it will use the DM-MP
device names such as

/dev/mapper/3600a098000aad73f00000a3f5a275dc8-part1

This is good if the system always runs in multipath mode. But if the system is
started without multipathing, booting with such an initramfs will fail,
because the /dev/mapper devices will not exist. The same problem can happen
with multipath maps and cloned SAN LUNs.

To prevent this from happening, the dracut policy for addressing disks
and partitions is changed to use /dev/disk/by-uuid device names on all
architectures except s390/s390x, which must be by-path (bsc#915218).

jsc#PED-1885