feat(nvmf): add code for parsing the NBFT
Add code to parse the NvmeoF Boot Firmware Table (NBFT).
The table is read by nvme-cli and printed in json format.
The jq utility is used for parsing the json output.

Set rd.neednet if valid nbft entries found; this is possible
because this script runs before 98-parse-ip-options.sh

Also add has_ibft_config markers for NBFT interfaces. We can use
"ibft" for this. Otherwise we'd have to modify the networking
tools. For the logic this is used for, all that matters is that
this is an external configuration mechanism applied during boot.

Usage of DHCP (either for obtaining initiator parameters and possibly
root-path) is indicated in "nvme show-nbft" output by a non-empty value
for dhcp_server_ipaddr in the HFI record.

jsc#PED-1449

(cherry picked from commit 870e3933012868005f2629489315fa3fe32e8db6)